move reducer logic to a slice that is scoped to the counter feature
