+++ src/features/AppHeader.js
@@ -11,7 +11,7 @@ const AppHeader = ({count}) => {
 
 function mapStateToProps(state) {
   return {
-    count: state.count
+    count: state.counter.value
   }
 }
